@@ -1,116 +1,103 @@
-// @ts-nocheck
+import CircularDependencyPlugin from "circular-dependency-plugin";
+import { resolve } from "path/posix";
+import webpack from "webpack";
+import { getAllResourceImages, getRevision, getVersion } from "./buildutils.js";
 
-const path = require("path");
-const webpack = require("webpack");
-const { getRevision, getVersion, getAllResourceImages } = require("./buildutils");
-const CircularDependencyPlugin = require("circular-dependency-plugin");
+const globalDefs = {
+    assert: "window.assert",
+    assertAlways: "window.assert",
+    abstract:
+        "window.assert(false, 'abstract method called of: ' + " +
+        "(this.name || (this.constructor && this.constructor.name)));",
+    G_IS_DEV: "true",
+    G_APP_ENVIRONMENT: JSON.stringify("development"),
+    G_BUILD_TIME: new Date().getTime().toString(),
+    G_BUILD_COMMIT_HASH: JSON.stringify(getRevision()),
+    G_BUILD_VERSION: JSON.stringify(getVersion()),
+    G_ALL_UI_IMAGES: JSON.stringify(getAllResourceImages()),
 
-module.exports = ({ watch = false, standalone = false }) => {
-    return {
-        mode: "development",
-        devtool: "cheap-source-map",
-        entry: {
-            "bundle.js": [path.resolve(__dirname, "../src/js/main.js")],
-        },
-        watch,
-        node: {
-            fs: "empty",
-        },
-        resolve: {
-            alias: {
-                "global-compression": path.resolve(__dirname, "..", "src", "js", "core", "lzstring.js"),
+    G_CHINA_VERSION: "false",
+    G_WEGAME_VERSION: "false",
+    G_GOG_VERSION: "false",
+    G_IS_RELEASE: "false",
+    G_IS_STANDALONE: "true",
+    G_IS_STEAM_DEMO: "false",
+    G_IS_BROWSER: "false",
+    G_HAVE_ASSERT: "true",
+};
+
+/** @type {import("webpack").RuleSetRule[]} */
+const moduleRules = [
+    {
+        test: /\.json$/,
+        enforce: "pre",
+        use: resolve("./loader.compressjson.cjs"),
+        type: "javascript/auto",
+    },
+    {
+        test: /\.js$/,
+        enforce: "pre",
+        exclude: /node_modules/,
+        use: [
+            {
+                loader: "webpack-strip-block",
+                options: {
+                    start: "typehints:start",
+                    end: "typehints:end",
+                },
             },
-        },
-        context: path.resolve(__dirname, ".."),
-        plugins: [
-            new webpack.DefinePlugin({
-                assert: "window.assert",
-                assertAlways: "window.assert",
-                abstract:
-                    "window.assert(false, 'abstract method called of: ' + (this.name || (this.constructor && this.constructor.name)));",
-                G_HAVE_ASSERT: "true",
-                G_APP_ENVIRONMENT: JSON.stringify("dev"),
-                G_IS_DEV: "true",
-                G_IS_RELEASE: "false",
-                G_IS_BROWSER: "true",
-                G_IS_STANDALONE: JSON.stringify(standalone),
-                G_BUILD_TIME: "" + new Date().getTime(),
-                G_BUILD_COMMIT_HASH: JSON.stringify(getRevision()),
-                G_BUILD_VERSION: JSON.stringify(getVersion()),
-                G_ALL_UI_IMAGES: JSON.stringify(getAllResourceImages()),
-            }),
-
-            new CircularDependencyPlugin({
-                // exclude detection of files based on a RegExp
-                exclude: /node_modules/,
-
-                // add errors to webpack instead of warnings
-                failOnError: true,
-
-                // allow import cycles that include an asyncronous import,
-                // e.g. via import(/* webpackMode: "weak" */ './file.js')
-                allowAsyncCycles: false,
-
-                // set the current working directory for displaying module paths
-                cwd: path.join(__dirname, "..", "src", "js"),
-            }),
         ],
-        module: {
-            rules: [
-                {
-                    test: /\.json$/,
-                    enforce: "pre",
-                    use: ["./gulp/loader.compressjson"],
-                    type: "javascript/auto",
+    },
+    {
+        test: /\.worker\.js$/,
+        use: [
+            {
+                loader: "worker-loader",
+                options: {
+                    filename: "[fullhash].worker.js",
+                    inline: "fallback",
                 },
-                { test: /\.(png|jpe?g|svg)$/, loader: "ignore-loader" },
-                { test: /\.nobuild/, loader: "ignore-loader" },
-                {
-                    test: /\.md$/,
-                    use: [
-                        {
-                            loader: "html-loader",
-                        },
-                        "markdown-loader",
-                    ],
-                },
-                {
-                    test: /\.js$/,
-                    enforce: "pre",
-                    exclude: /node_modules/,
-                    use: [
-                        {
-                            loader: "webpack-strip-block",
-                            options: {
-                                start: "typehints:start",
-                                end: "typehints:end",
-                            },
-                        },
-                        {
-                            loader: path.resolve(__dirname, "mod.js"),
-                        },
-                    ],
-                },
-                {
-                    test: /\.worker\.js$/,
-                    use: {
-                        loader: "worker-loader",
-                        options: {
-                            fallback: false,
-                            inline: true,
-                        },
-                    },
-                },
-                {
-                    test: /\.ya?ml$/,
-                    type: "json", // Required by Webpack v4
-                    use: "yaml-loader",
-                },
-            ],
+            },
+        ],
+    },
+    {
+        test: /\.js$/,
+        resolve: {
+            fullySpecified: false,
         },
-        output: {
-            filename: "bundle.js",
-            path: path.resolve(__dirname, "..", "build"),
+    },
+];
+
+/** @type {import("webpack").Configuration} */
+export default {
+    mode: "development",
+    entry: resolve("../src/js/main.js"),
+    context: resolve(".."),
+    output: {
+        path: resolve("../build"),
+        filename: "bundle.js",
+    },
+    resolve: {
+        fallback: { fs: false },
+        alias: {
+            "global-compression": resolve("../src/js/core/lzstring.js"),
         },
-    };
+    },
+    devtool: "cheap-source-map",
+    watch: true,
+    plugins: [
+        new webpack.DefinePlugin(globalDefs),
+        new webpack.IgnorePlugin({ resourceRegExp: /\.(png|jpe?g|svg)$/ }),
+        new webpack.IgnorePlugin({ resourceRegExp: /\.nobuild/ }),
+        new CircularDependencyPlugin({
+            exclude: /node_modules/,
+            failOnError: true,
+            allowAsyncCycles: false,
+            cwd: resolve("../src/js"),
+        }),
+    ],
+    module: { rules: moduleRules },
+    experiments: {
+        topLevelAwait: true,
+    },
 };
